+++ NEWS.md
@@ -1,6 +1,8 @@
 # Version 0.0.2
 
 Renamed package shuffle, expanded functionality to stratified
+- created random assignment functions
+- added unit testing for mahalanobis distance  
 
 # Version 0.0.1